+++ lib/vagrant/box_metadata.rb
@@ -31,12 +31,14 @@ module Vagrant
       @name = @raw["name"]
       @description = @raw["description"]
       @version_map = (@raw["versions"] || []).map do |v|
-        [Gem::Version.new(v["version"]), v]
+        begin
+          [Gem::Version.new(v["version"]), v]
+        rescue ArgumentError
+          raise Errors::BoxMetadataMalformedVersion,
+            version: v["version"].to_s
+        end
       end
       @version_map = Hash[@version_map]
-
-      # TODO: check for corruption:
-      #   - malformed version
     end
 
     # Returns data about a single version that is included in this